# packages/contracts/openapi/xtiitch.v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xtiitch API",
    "version": "0.1.0",
    "description": "Seed REST contract for the active Xtiitch v1 public storefront, business dashboard, admin console, money, growth, and webhook route families."
  },
  "servers": [
    {
      "url": "https://api.xtiitch.com",
      "description": "Production API"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development API"
    }
  ],
...
  "paths": {
...
    "/v1/businesses/me": {
      "get": {
        "operationId": "getBusinessProfile",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "responses": { "200": { "$ref": "#/components/responses/JsonOk" } }
      }
    },
...
    "/v1/store-settings": {
      "get": {
        "operationId": "getStoreSettings",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "responses": { "200": { "$ref": "#/components/responses/JsonOk" } }
      },
      "patch": {
        "operationId": "updateStoreSettings",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
        "responses": { "200": { "$ref": "#/components/responses/JsonOk" } }
      }
    },
    "/v1/collections": {
      "get": {
        "operationId": "listCollections",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "responses": { "200": { "$ref": "#/components/responses/JsonOk" } }
      },
      "post": {
        "operationId": "createCollection",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
        "responses": { "201": { "$ref": "#/components/responses/JsonOk" } }
      }
    },
    "/v1/designs": {
      "get": {
        "operationId": "listDesigns",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "responses": { "200": { "$ref": "#/components/responses/JsonOk" } }
      },
      "post": {
        "operationId": "createDesign",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
        "responses": { "201": { "$ref": "#/components/responses/JsonOk" } }
      }
    },
    "/v1/designs/{id}/prices/{bandId}": {
      "put": {
        "operationId": "setDesignPrice",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Id" },
          { "$ref": "#/components/parameters/BandId" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
        "responses": { "200": { "$ref": "#/components/responses/JsonOk" } }
      }
    },
...
    "/v1/promotions": {
      "get": {
        "operationId": "listBusinessPromotions",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "responses": { "200": { "$ref": "#/components/responses/JsonOk" } }
      },
      "post": {
        "operationId": "createBusinessPromotion",
        "tags": ["Business dashboard"],
        "security": [{ "BusinessBearer": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/JsonBody" },
        "responses": { "201": { "$ref": "#/components/responses/JsonOk" } }
      }
    },
    "/v1/public/stores/{handle}": {
      "get": {
        "operationId": "getPublicStore",
        "tags": ["Public storefront"],
        "parameters": [{ "$ref": "#/components/parameters/Handle" }],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/public/designs/{handle}": {
      "get": {
        "operationId": "getPublicDesign",
        "tags": ["Public storefront"],
        "parameters": [{ "$ref": "#/components/parameters/Handle" }],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BusinessBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Business dashboard access token with tenant scope derived server-side."
      },
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Platform admin access token with operator role and permissions."
      },
      "CustomerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Customer session token issued by the customer OTP sign-in; scopes the caller to their own orders."
      },
      "PaystackSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Paystack-Signature",
        "description": "HMAC-SHA512 signature over the raw Paystack webhook body."
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "BandId": {
        "name": "bandId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "Handle": {
        "name": "handle",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9-]*[a-z0-9]$"
        }
      },
...
    },
    "requestBodies": {
      "JsonBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/GenericJson" }
          }
        }
      }
    },
    "responses": {
      "JsonOk": {
        "description": "JSON response.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/GenericJson" }
          }
        }
      },
...
      "Error": {
        "description": "Error response.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "GenericJson": {
        "type": "object",
        "additionalProperties": true
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "additionalProperties": true
      }
    }
  }
}